Add ingress-nginx chart to helm & Fix next.js envs
