+++ .gitignore
@@ -35,7 +35,6 @@ coverage
 !.vscode/launch.json
 !.vscode/extensions.json
 
-
 ## NextJS
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
