+++ application/services/web/.env.example
@@ -1,3 +1,2 @@
 NEXT_PUBLIC_API_HOST=aloco.local
-API_URL=aloco.local
-NODE_ENV=local
+NEXT_PUBLIC_ENV=local
