--- application/services/web/src/common/utils/constants.ts
@@ -1,1 +1,0 @@
-export const ENV = process.env.NODE_ENV
