+++ application/services/web/src/common/utils/envs.ts
@@ -1,0 +1,15 @@
+enum Environment {
+  PRODUCTION = 'production',
+  DEVELOPMENT = 'development',
+  LOCAL = 'local',
+}
+
+const ENV = process.env.NEXT_PUBLIC_ENV ?? '__REPLACE__NEXT_PUBLIC_ENV__'
+
+export const env = () => ENV
+
+export const production = () => ENV === Environment.PRODUCTION
+
+export const development = () => ENV === Environment.DEVELOPMENT
+
+export const local = () => ENV === Environment.LOCAL
