--- application/services/web/src/network/common/constants.ts
@@ -1,3 +1,0 @@
-export const API_HOST = process.env.NEXT_PUBLIC_API_HOST
-export const INGRESS_API_URL =
-  'http://ingress-nginx-controller.ingress-nginx.svc.cluster.local'
