+++ application/services/web/src/ui/common/utils/get-headers.ts
@@ -1,7 +1,7 @@
 import {headers as nextHeaders} from 'next/dist/client/components/headers'
 
-import {local} from '@/common/utils/env'
-import {API_HOST} from '@/network/common/constants'
+import {local} from '@/common/utils/envs'
+import {API_HOST} from '@/network/common/envs'
 
 export const getHeaders = () => {
   const headers = Object.fromEntries(nextHeaders().entries())
