+++ infrastructure/images/web/cmd.sh
@@ -1,7 +1,8 @@
 cmd="$1"
 
 image="dormammun/looport-web:latest"
-arch="linux/arm64,linux/amd64"
+arch="linux/amd64"
+#arch="linux/arm64,linux/amd64"
 
 case $cmd in
 "run")
